Backend/requirements.txt
asgiref==3.11.1
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
bcrypt==5.0.0
dnspython==2.8.0
mongoengine==0.29.1
Pillow==10.1.0
pymongo==4.16.0
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.5
tzdata==2025.3
sendgrid==6.12.5
django-sendgrid-v5==1.3.1
